# storage/api-docs/openapi.yaml
openapi: 3.0.3
info:
  title: 2B Academy LMS API
  description: |
    RESTful API for the 2B Academy Learning Management System.
    All endpoints are prefixed with `/api/v1`.

    **Authentication:** Protected endpoints require a Bearer token obtained from the login endpoints.
    Include it in all requests as: `Authorization: Bearer <token>`

    **Response format:**
    ```json
    { "status": "success|error", "message": "...", "result": {} }
    ```
    Paginated responses also include a `meta` key:
    ```json
    { "meta": { "current_page": 1, "last_page": 5, "per_page": 20, "total": 100 } }
    ```

    **Translatable fields** return bilingual objects:
    ```json
    { "ar": "اسم عربي", "en": "English name" }
    ```
  version: 1.0.0
  contact:
    name: 2B Academy API
servers:
  - url: /api/v1
    description: API v1

# ---------------------------------------------------------------------------
# Security
# ---------------------------------------------------------------------------
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Sanctum

  # -------------------------------------------------------------------------
  # Reusable schemas
  # -------------------------------------------------------------------------
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        status: { type: string, example: success }
        message: { type: string }
        result: {}
...
    ErrorResponse:
      type: object
      properties:
        status: { type: string, example: error }
        error:  { type: string }
    ValidationErrorResponse:
      type: object
      properties:
        message: { type: string }
        errors:
          type: object
          additionalProperties:
            type: array
            items: { type: string }

    # -------------------------------------------------------------------------
    # Domain schemas
    # -------------------------------------------------------------------------
    User:
      type: object
      properties:
        id:              { type: integer }
        name:            { type: string }
        email:           { type: string, format: email }
        phone:           { type: string, nullable: true }
        system_id:       { type: string, nullable: true }
        machine_code:    { type: string, nullable: true }
        department_name: { type: string, nullable: true }
        roles:           { type: array, items: { type: string } }
        created_at:      { type: string, format: date-time }
    Admin:
      type: object
      properties:
        id:         { type: integer }
        name:       { type: string }
        email:      { type: string, format: email }
        roles:      { type: array, items: { type: string } }
        created_at: { type: string, format: date-time }
...
  responses:
    Unauthorized:
      description: Missing or invalid token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: Insufficient role
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ValidationErrorResponse' }

# =============================================================================
# Tags
# =============================================================================
...
paths:
...
  /auth/user/login:
    post:
      tags: [Auth]
      summary: User login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:    { type: string, format: email }
                password: { type: string, minLength: 8 }
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      result:
                        type: object
                        properties:
                          token: { type: string }
                          user:  { $ref: '#/components/schemas/User' }
        '401': { description: Invalid credentials }
        '422': { $ref: '#/components/responses/ValidationError' }
  /auth/user/logout:
    post:
      tags: [Auth]
      summary: Logout current user session
      security: [{ BearerAuth: [] }]
      responses:
        '200': { description: Logged out }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/user/logout-all:
    post:
      tags: [Auth]
      summary: Logout all user sessions
      security: [{ BearerAuth: [] }]
      responses:
        '200': { description: All sessions invalidated }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/user/me:
    get:
      tags: [Auth]
      summary: Get authenticated user profile
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      result: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/admin/login:
    post:
      tags: [Auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:    { type: string, format: email }
                password: { type: string }
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      result:
                        type: object
                        properties:
                          token: { type: string }
                          admin: { $ref: '#/components/schemas/Admin' }
        '401': { description: Invalid credentials }
  /auth/admin/logout:
    post:
      tags: [Auth]
      summary: Admin logout
      security: [{ BearerAuth: [] }]
      responses:
        '200': { description: Logged out }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/admin/me:
    get:
      tags: [Auth]
      summary: Get authenticated admin profile
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: Admin profile
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      result: { $ref: '#/components/schemas/Admin' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }

  # ===========================================================================
  # CATEGORIES
  # ===========================================================================
...